[Refactor] RestAPI에 맞게 URI 수정
