+++ models/comment.js
@@ -4,14 +4,13 @@ const createComment = async (commentInfo) => {
   const newComment = await Comment.create(commentInfo);
   return newComment;
 };
-const findComment = async () => {};
+
 const findComments = async () => {};
 const updateComment = async () => {};
 const destroyComment = async () => {};
 
 module.exports = {
   createComment,
-  findComment,
   findComments,
   updateComment,
   destroyComment,
